+++ packages/react-client-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @paperxyz/react-client-sdk
 
+## 1.1.0
+
+### Minor Changes
+
+- 48a09b8: fix: hide loading spinner if no callback from inner iframe takes too long
+
+## 1.0.4
+
+### Patch Changes
+
+- 3f8dffc: iframe now takes 100% height (w/ minHeight)
+- Updated dependencies [3f8dffc]
+  - @paperxyz/js-client-sdk@0.2.2
+
 ## 1.0.3
 
 ### Patch Changes